third_party/blink/web_tests/accessibility/dl-role.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<dl id="dlElem">
    <dt id="dtElem">term</dt>
    <dd id="ddElem">definition</dd>
</dl>

<script>
description("This tests that AXRole for dl element is AXDescriptionList.")

if (window.accessibilityController) {
    var dlElement = accessibilityController.accessibleElementById("dlElem");
    shouldBe("dlElement.role", "'AXRole: AXDescriptionList'");

    var dtElement = accessibilityController.accessibleElementById("dtElem");
    shouldBe("dtElement.role", "'AXRole: AXTerm'");

    var ddElement = accessibilityController.accessibleElementById("ddElem");
    shouldBe("ddElement.role", "'AXRole: AXDefinition'");
}
</script>
