third_party/blink/web_tests/accessibility/dl-role-expected.txt

This tests that AXRole for dl element is AXDescriptionList.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dlElement.role is 'AXRole: AXDescriptionList'
PASS dtElement.role is 'AXRole: AXTerm'
PASS ddElement.role is 'AXRole: AXDefinition'
PASS successfullyParsed is true

TEST COMPLETE

term
definition
